+++ packages/admin-ui/server/AdminUI.js
@@ -1,14 +1,15 @@
-const falsey = require('falsey');
 const express = require('express');
 const webpack = require('webpack');
 const chalk = require('chalk');
 const terminalLink = require('terminal-link');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const webpackHotMiddleware = require('webpack-hot-middleware');
+const compression = require('compression');
+const { createSessionMiddleware } = require('@keystone-alpha/session');
+
 const pkgInfo = require('../package.json');
 
 const getWebpackConfig = require('./getWebpackConfig');
-const { createSessionMiddleware } = require('./sessionMiddleware');
 const { mode } = require('./env');
 
 module.exports = class AdminUI {
@@ -36,13 +37,13 @@ module.exports = class AdminUI {
 
   getAdminMeta() {
     return {
-      withAuth: !!this.authStrategy,
-      authList: this.authStrategy ? this.authStrategy.listKey : null,
       adminPath: this.adminPath,
+      authList: this.authStrategy ? this.authStrategy.listKey : null,
+      pages: this.config.pages,
+      sessionPath: this.config.sessionPath,
       signinPath: this.config.signinPath,
       signoutPath: this.config.signoutPath,
-      sessionPath: this.config.sessionPath,
-      sortListsAlphabetically: this.config.sortListsAlphabetically,
+      withAuth: !!this.authStrategy,
     };
   }
 
@@ -50,7 +51,6 @@ module.exports = class AdminUI {
     const { signinPath, signoutPath, sessionPath } = this.config;
     return createSessionMiddleware(
       { signinPath, signoutPath, sessionPath, successPath: this.adminPath },
-      this.keystone.sessionManager,
       this.authStrategy
     );
   }
@@ -60,13 +60,20 @@ module.exports = class AdminUI {
     const { adminPath } = this;
 
     // ensure any non-resource requests are rewritten for history api fallback
-    if (falsey(process.env.DISABLE_LOGGING)) {
+    if (process.env.NODE_ENV !== 'production') {
       const url = `http://localhost:${port}${adminPath}`;
       const prettyUrl = chalk.blue(`${url}(/.*)?`);
       const clickableUrl = terminalLink(prettyUrl, url, { fallback: () => prettyUrl });
 
       console.log(`🔗 ${chalk.green('Keystone Admin UI:')} ${clickableUrl} (v${pkgInfo.version})`);
     }
+
+    if (mode === 'production') {
+      // only use compression in production because it breaks server sent events
+      // which is what webpack-hot-middleware uses
+      app.use(compression());
+    }
+
     app.use(adminPath, (req, res, next) => {
       // TODO: make sure that this change is OK. (regex was testing on url, not path)
       // Changed because this was preventing adminui pages loading when a querystrings
@@ -143,7 +150,6 @@ module.exports = class AdminUI {
         });
       };
     }
-
     // handle errors
     // eslint-disable-next-line no-unused-vars
     app.use(function(err, req, res, next) {