+++ test-projects/access-control/server.js
@@ -26,8 +26,6 @@ keystone
       });
       await keystoneApp.createItems(initialData);
     }
-
-    console.log(`Listening on port ${port}`);
   })
   .catch(error => {
     console.error(error);