+++ test-projects/twitter-login/twitter.js
@@ -1,5 +1,5 @@
 const TwitterAuthStrategy = require('@keystone-alpha/keystone/auth/Twitter');
-
+const { startAuthedSession } = require('@keystone-alpha/session');
 const { appURL, twitterAppKey, twitterAppSecret } = require('./config');
 
 exports.configureTwitterAuth = function(keystone, server) {
@@ -27,6 +27,14 @@ exports.configureTwitterAuth = function(keystone, server) {
       },
     })
   );
+
+  server.app.get('/api/session', (req, res) => {
+    res.json({
+      signedIn: !!req.session.keystoneItemId,
+      userId: req.session.keystoneItemId,
+      name: req.user ? req.user.name : undefined,
+    });
+  });
 
   // Twitter will redirect the user to this URL after approval.
   server.app.get(
@@ -98,7 +106,7 @@ exports.configureTwitterAuth = function(keystone, server) {
         });
 
         await keystone.auth.User.twitter.connectItem(req, { item });
-        await keystone.sessionManager.startAuthedSession(req, { item, list });
+        await startAuthedSession(req, { item, list });
         res.redirect('/api/session');
       } catch (createError) {
         next(createError);