+++ packages/arch/packages/navbar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @arch-ui/navbar
 
+## 0.0.4
+
+- [patch][85cb44a9](https://github.com/keystonejs/keystone-5/commit/85cb44a9):
+
+  - Add PrimaryNavHeading component and change styles
+
 ## 0.0.3
 
 - [patch][7417ea3a](https://github.com/keystonejs/keystone-5/commit/7417ea3a):